requirements.txt
asgiref==3.8.1
attr==0.3.2
Babel==2.16.0
ConfigParser==7.1.0
contextlib2==21.6.0
cryptography==43.0.1
docutils==0.21.2
email_validator==2.2.0
filelock==3.15.4
flask_babel==4.0.0
Flask_Cors==4.0.1
Flask_Login==0.6.3
flask_sqlalchemy==3.1.1
HTMLParser==0.0.2
importlib_metadata==8.4.0
ipython==8.12.3
ipywidgets==8.1.5
jnius==1.1.0
keyring==25.3.0
mock==5.1.0
Pillow==10.4.0
protobuf==5.28.0
pyOpenSSL==24.2.1
python-dotenv==1.0.1
railroad==0.5.0
redis==5.0.8
Sphinx==8.0.2
thread==2.0.5
tornado==6.4.1
urllib3_secure_extra==0.1.0
watchdog==5.0.2
xmlrpclib==1.0.1
